+++ backend/packages/server/src/app.ts
@@ -38,6 +38,7 @@ import { validationConfigController } from './modules/validation/ValidationConfi
 import { packageRegistriesController } from './modules/packageRegistries/PackageRegistriesController.js'
 import { previewController } from './modules/preview/PreviewController.js'
 import { incidentEnrichmentController } from './modules/incidentEnrichment/IncidentEnrichmentController.js'
+import { agentPromptController } from './modules/agentPrompts/AgentPromptController.js'
 import { modelPresetController } from './modules/modelPresets/ModelPresetController.js'
 import { serviceFragmentDefaultsController } from './modules/serviceFragmentDefaults/ServiceFragmentDefaultsController.js'
 import { modelController } from './modules/models/ModelController.js'
@@ -274,6 +275,7 @@ function registerWorkspaceConfigControllers<E extends AppEnv>(app: Hono<E>): voi
   app.route('/workspaces/:workspaceId', previewController())
   app.route('/workspaces/:workspaceId', incidentEnrichmentController())
   app.route('/workspaces/:workspaceId', modelPresetController())
+  app.route('/workspaces/:workspaceId', agentPromptController())
   app.route('/workspaces/:workspaceId', serviceFragmentDefaultsController())
   app.route('/workspaces/:workspaceId', recurringPipelineController())
   app.route('/workspaces/:workspaceId', trackerSettingsController())