+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cat-factory/provider-s3
 
+## 0.2.240
+
+### Patch Changes
+
+- Updated dependencies [e087b40]
+  - @cat-factory/kernel@0.186.0
+
 ## 0.2.239
 
 ### Patch Changes