+++ backend/packages/spend/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cat-factory/spend
 
+## 0.12.116
+
+### Patch Changes
+
+- Updated dependencies [e087b40]
+  - @cat-factory/contracts@0.191.0
+  - @cat-factory/kernel@0.186.0
+
 ## 0.12.115
 
 ### Patch Changes